added validation in SP

+++ public/deletecustomer.sql
@@ -3,10 +3,18 @@ create procedure deletecustomer(INOUT cuscode integer, INOUT addressid integer D
 as
 $$
     BEGIN
-        addressId = (SELECT CAddressId FROM CustomerAddress WHERE CustomerCode=cusCode);
-        DELETE FROM CustomerContact WHERE CAddressId = addressId;
-        DELETE FROM CustomerAddress WHERE CustomerCode = cusCode;
-        DELETE FROM Customers WHERE CustomerCode = cusCode;
+        IF EXISTS(SELECT CustomerCode FROM customers WHERE customercode=cusCode)
+            THEN
+                IF NOT EXISTS(SELECT customercode FROM salestransaction WHERE customercode=cusCode)
+                    THEN
+                        addressId = (SELECT CAddressId FROM CustomerAddress WHERE CustomerCode=cusCode);
+                        DELETE FROM CustomerContact WHERE CAddressId = addressId;
+                        DELETE FROM CustomerAddress WHERE CustomerCode = cusCode;
+                        DELETE FROM Customers WHERE CustomerCode = cusCode;
+                    ELSE RAISE 'Cannot delete customer with transactions.';
+                end if;
+            ELSE RAISE 'Customer does not exists.';
+        END IF;
         COMMIT;
     END
 $$;
